cli: data table
Include the table component. It's used to show the data fetched from API.
The table is composed by a command line showing which command is
available and the table itself.

The user can use up arrow and down arrow to navigates in the table rows.
In the future some commands from command line will considered the select
row to take action